created different files for db, handlers, env, keyboards
