+++ Polina_DOD/.gitignore
@@ -1,0 +1,2 @@
+/.vscode
+/__pycache__
